.pack_and_function_check

Check if 1) required functions are present in required packages and 2) required packages are installed locally. Simplified version of saferDev::is_function_here(), used as internal function for the other functions of the package.

Usage

.pack_and_function_check(fun, lib_path, error_text)

Arguments

fun

Character vector of the names of the required functions, preceded by the name of the package they belong to and a double or triple colon. Example: c("ggplot2::geom_point", "grid::gpar").

lib_path

Vector of characters specifying the absolute pathways of the directories containing the required packages for the function, if not in the default directories. Useful when R package are not installed in the default directories because of lack of admin rights. More precisely, lib_path is passed through the new argument of .libPaths() so that the new library paths are unique(c(new, .Library.site, .Library)). Warning: .libPaths() is restored to the initial paths, after function execution. Ignored if NULL (default) or if the safer_check argument is FALSE: only the pathways specified by the current .libPaths() are used for package calling.

error_text

Single character string used to add information in error messages returned by the function, notably if the function is inside other functions, which is practical for debugging. Example: error_text = " INSIDE <PACKAGE_1>::<FUNCTION_1> INSIDE <PACKAGE_2>::<FUNCTION_2>.". If NULL, converted into "".

Value

An error message if at least one of the checked packages is missing in lib_path, or if at least one of the checked functions is missing in the required package, nothing otherwise.

Author

Gael Millot

Yushi Han

Haiding Wang

Gael Millot <gael.millot@pasteur.fr>

Examples

if (FALSE)  # Example that shouldn't be run because this is an internal function (not found by devtools::check())
saferDev:::.pack_and_function_check(fun = 1, lib_path = NULL, error_text = " INSIDE F1.") # this example returns an error
saferDev:::.pack_and_function_check(fun = "ggplot2::notgood", lib_path = base::.libPaths(), error_text = " INSIDE P1::F1") # this example returns an error
#> Error: 
#> 
#> ================
#> 
#> ERROR IN saferDev:::.pack_and_function_check() INSIDE P1::F1
#> 
#> REQUIRED FUNCTION IS MISSING IN THE INSTALLED PACKAGE:
#> ggplot2::notgood
#> 
#> IN ONE OF THESE FOLDERS:
#> /__w/_temp/Library
#> /usr/local/lib/R/site-library
#> /usr/local/lib/R/library
#> 
#> ================
#> 
saferDev:::.pack_and_function_check(fun = c("ggplot2::geom_point", "grid::gpar"), lib_path = base::.libPaths(), error_text = " INSIDE P1::F1")  # nothing should happen